CRUD endereco, pessoa, perfil, pessoaPerfil

+++ src/main/java/com/stefanini/servico/PessoaServico.java
@@ -27,51 +27,45 @@ import com.stefanini.model.Pessoa;
 @TransactionAttribute(TransactionAttributeType.NOT_SUPPORTED)
 public class PessoaServico implements Serializable {
 
-	/**
-	 * 
-	 */
+
 	private static final long serialVersionUID = 1L;
+	
+	
 	@Inject
 	private PessoaDAO dao;
+	@Inject
+	private EnderecoServico enderecoService;
 
-	/**
-	 * Salvar os dados de uma Pessoa
-	 */
 	@TransactionAttribute(TransactionAttributeType.REQUIRES_NEW)
-	public Pessoa salvar(@Valid Pessoa pessoa) {
-		return dao.salvar(pessoa);
+	public Pessoa salvar(@Valid Pessoa pessoa) throws Exception {
+		if(dao.buscarEmailPessoa(pessoa.getEmail()).get().isEmpty()) {
+			return dao.salvar(pessoa);
+		}
+		throw new Exception("Este email já existe");
 	}
 
-	/**
-	 * Atualizar o dados de uma pessoa
-	 */
-//	@Override
+
 	@TransactionAttribute(TransactionAttributeType.REQUIRES_NEW)
 	public Pessoa atualizar(@Valid Pessoa entity) {
 		return dao.atualizar(entity);
 	}
 
-	/**
-	 * Remover uma pessoa pelo id
-	 */
-//	@Override
+
 	@TransactionAttribute(TransactionAttributeType.REQUIRES_NEW)
-	public void remover(@Valid Long id) {
-		dao.remover(id);
+	public void remover(@Valid Long id) throws Exception {
+		if(enderecoService.buscarSePessoaPossuiEndereco(id).get().isEmpty()) {
+			dao.remover(id);	
+		} else {
+			throw new Exception("Usuario possui endereço vinculado!");			
+		}
 	}
 
-	/**
-	 * Buscar uma lista de Pessoa
-	 */
-//	@Override
+
 	public Optional<List<Pessoa>> getList() {
 		return dao.getList();
 	}
+	
 
-	/**
-	 * Buscar uma Pessoa pelo ID
-	 */
-//	@Override
 	public Optional<Pessoa> encontrar(Long id) {
 		return dao.encontrar(id);
 	}
